+++ 3574/1/output.rb
@@ -1,6 +1,8 @@
 require 'terminal-table'
 
 class Output
+  HEADER = %i[name_id used_by watch stars fork issues contributors].freeze
+
   def call(data)
     rows = create_rows(data)
     table = create_table(data, rows)
